+++ rakaia/callbacks/roi.py
@@ -77,6 +77,7 @@ def init_roi_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                        State('spatial-spot-rad', 'value'),
                        State('mask-in-gallery', 'value'),
                        State('query-min-obj', 'value'),
+                       State('subsample-roi-gallery', 'value'),
                        prevent_initial_call=True)
     @DownloadDirGenerator(os.path.join(tmpdirname, authentic_id, str(uuid.uuid1()), 'downloads'))
     def generate_roi_images_from_query(export_roi, currently_selected, data_selection, blend_colour_dict,
@@ -85,7 +86,7 @@ def init_roi_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
                                     dataset_options, query_cell_id_lists, global_apply_filter,
                                     global_filter_type, global_filter_val, global_filter_sigma, delimiter, error_config,
                                     dim_min, dim_max, keyw, saved_blend, saved_blend_dict, find_similar, image_cor, quant,
-                                    spatial_rad, enable_masks, query_min):
+                                    spatial_rad, enable_masks, query_min, subsample_thumbnail):
         """
         Generate the dynamic gallery of ROI queries from the query selection
         Can be activated using either the original button for a fresh query, or the button to load additional ROIs
@@ -121,7 +122,7 @@ def init_roi_level_callbacks(dash_app, tmpdirname, authentic_id, app_config):
             mask_dict, dataset_options, query_cell_id_lists, global_apply_filter, global_filter_type, global_filter_val, global_filter_sigma,
             delimiter, False, dim_min, dim_max, keyw, False, spatial_rad, enable_masks, True,
                     app_config['array_store_type'], query_min).get_image_dict()
-            new_row_children, roi_list = roi_query_gallery_children(images)
+            new_row_children, roi_list = roi_query_gallery_children(images, subsample_thumbnail=subsample_thumbnail)
             # if the query is being extended, append to the existing gallery for exclusion. Otherwise, start fresh
             if ctx.triggered_id == "dataset-query-additional-load": roi_list = list(set(rois_exclude + roi_list))
             roi_list.append(data_selection)